clausewise/requirements.txt
gradio>=4.44
transformers>=4.44.2
accelerate>=0.34.2
torch
pandas>=2.0
pypdf>=3.0
python-docx>=1.1.0
spacy>=3.8.0
sentencepiece
safetensors
vaderSentiment
peft>=0.12.0
